Added options contentType, encodeValue, full POST json support and support URL as value
,

Added the contentType ajax option defaults to
'application/x-www-form-urlencoded' can be changed to 'application/json;
charset=utf-8' which is required for .net web service support. To make
this work the response must be changed for data.d - this has been added.

Also to support the use of a URL as the returned value the
encodeURIComponent in the buildItems function must be able disabled, so
the option encodeValue was added this defaults to true which means the
value is encoded - set to false to support URL or non encode required
values.